# packages/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: aos-api
  version: 0.3.0
  description: >
    Stable contract surface per docs/palantier/20_tech/T-API-aos-api稳定契约.md.
    Aligned Wave-0～5 + G-ALIGN (47) — main paths listed; see T-API for full narrative.
servers:
  - url: /v1
paths:
...
  /authz/status:
    get:
      operationId: authzStatus
      summary: OpenFGA facade status (local vs remote sidecar · schemes 58/61)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: mode · reachable · storeId · modelVersion
        "401":
          $ref: "#/components/responses/Unauthorized"
  /authz/model:
    get:
      operationId: authzModel
      summary: OpenFGA type/relation catalog (scheme 61 aos-prod-v1)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: modelVersion · types · relations
        "401":
          $ref: "#/components/responses/Unauthorized"
  /authz/check:
    post:
      operationId: authzCheck
      summary: OpenFGA-shaped authorization check (T-CROSS · schemes 55/58)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object]
              properties:
                user:
                  type: string
                  description: defaults to user:{sub}
                relation:
                  type: string
                  default: viewer
                object:
                  type: string
                  example: object:WorkOrder:wo-fga-demo
      responses:
        "200":
          description: allowed boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
  /authz/tuples:
    post:
      operationId: authzWriteTuple
      summary: Write authz tuple (local + optional OpenFGA Write)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object]
              properties:
                user:
                  type: string
                relation:
                  type: string
                  default: viewer
                object:
                  type: string
      responses:
        "200":
          description: ok
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /modules:
    get:
      operationId: listModules
      summary: Module list (T08)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
        "501":
          description: Wave-1
    post:
      operationId: createModule
      summary: Create Module
      security:
        - bearerAuth: []
      responses:
        "201":
          description: created
        "501":
          description: Wave-1
...
  /modules/{id}:
    get:
      operationId: getModule
      summary: Get Module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
        "403":
          description: FORBIDDEN
        "404":
          description: missing
    patch:
      operationId: patchModule
      summary: Patch Module definition
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
  /modules/{id}/publish:
    post:
      operationId: publishModule
      summary: Apollo Publish Adapter (Lite)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: published
  /modules/{id}/runtime:
    get:
      operationId: moduleRuntime
      summary: Module runtime schema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
  schemas:
    ErrorBody:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
...
  responses:
    Unauthorized:
      description: AUTH_REQUIRED
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"